# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario de Redes",
        "description": "Control de creación, actualización, visualización y eliminación de registros de usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/types": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "( Muestra todos los tipos de dispositivos registrados)",
                "operationId": "5ae5a6ccb7da5e1410da3b56b6fde09f",
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "anyOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Type1Schema"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Type2Schema"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/types/{id}": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "( Muestra datos de un tipo de dispositivo identificado por {id})",
                "operationId": "6bd0b47f25aa9325d8aa96c18ec401a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "router"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "( Actualiza los datos de un tipo de dispositivo identificado por {id})",
                "operationId": "4f384bc0a685f6a6c172bce5ad7fb6dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "number",
                                        "example": "router"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Datos de tipo de dispositivo actualizados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro actualizado exitosamente"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "router"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "( Elimina los datos de un tipo (type) identificado por {id})",
                "operationId": "103812d7190b22f2d47edcb4cf7322f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos de estado de dispositivo borrados exitosamente"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "switch"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/types/register": {
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "( Crea un nuevo tipo de dispositivo)",
                "operationId": "35117ba2ae9d7d0f763f67e4355e7b20",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "number",
                                        "example": "switch"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de dispositivo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro creado exitosamente"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "switch"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Type1Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "servidor blade"
                    }
                },
                "type": "object"
            },
            "Type2Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "2"
                    },
                    "name": {
                        "type": "number",
                        "example": "switch"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}